+++ tests/Support/FixtureData.php
@@ -86,21 +86,27 @@ final readonly class FixtureData implements DownloadResource
         }
     }
 
-    public function output(int $length, int $start = 0, int $bytesPerSecond = 0): void
+    public function output(DownloadConfig $config, int $length, int $start = 0): void
     {
         if ($length <= 0 || empty($this->data)) {
             return;
         }
 
-        if (($stream = fopen('php://temp', 'r+')) === false) {
+        if (($handle = fopen('php://temp', 'r+')) === false) {
             return;
         }
 
         try {
-            fwrite($stream, $this->data);
-            $this->outputStream($stream, $length, $start, $bytesPerSecond);
+            fwrite($handle, $this->data);
+            $this->outputStream(
+                $handle,
+                $length,
+                $start,
+                $config->getBytesPerSecond(),
+                $config->getChunkSize(),
+            );
         } finally {
-            fclose($stream);
+            fclose($handle);
         }
     }